refactor: 각 파일들 위치 정리
