+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./AccessableModal";
+export * from "./FormModalContent";
+export * from "./ModalManager";
