+++ src/main.tsx
@@ -2,7 +2,7 @@ import "modern-normalize";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import ModalFormPage from "./ModalFormPage";
+import { ModalFormPage } from "./pages";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
